requirements.txt
openai>=1.40.0
feedparser>=6.0.11
chromadb>=0.5.0
pydantic>=2.8.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
requests>=2.32.0